+++ src/generate.h
@@ -329,7 +329,7 @@ const char *get_level_profile_name_from_index(int i);
 struct chunk *town_gen(struct player *p, int min_height, int min_width,
 	const char **p_error);
 struct chunk *town_gen_wild(struct player *p, uint32_t seed,
-	int wid, int hgt, uint16_t shops);
+	int wid, int hgt, uint16_t shops, uint16_t services);
 struct chunk *classic_gen(struct player *p, int min_height, int min_width,
 	const char **p_error);
 struct chunk *labyrinth_gen(struct player *p, int min_height, int min_width,