+++ src/init.c
@@ -605,6 +605,8 @@ static enum parser_error parse_constants_wild(struct parser *p) {
 		z->wild_lakes = value;
 	} else if (streq(label, "road-dist")) {
 		z->wild_road_dist = value;
+	} else if (streq(label, "travel-cost")) {
+		z->wild_travel_cost = value;
 	} else if (streq(label, "gate-turns")) {
 		z->wild_gate_turns = value;
 	} else if (streq(label, "towns")) {
@@ -1205,6 +1207,8 @@ static errr finish_parse_constants(struct parser *p) {
 		z_info->wild_lakes = 4;
 	if (!z_info->wild_road_dist)
 		z_info->wild_road_dist = 30;
+	if (!z_info->wild_travel_cost)
+		z_info->wild_travel_cost = 15;
 	if (!z_info->wild_gate_turns)
 		z_info->wild_gate_turns = 100;
 	if (!z_info->wild_towns)