+++ src/savefile.h
@@ -100,6 +100,7 @@ int rd_stores(void);
 int rd_wilderness_1(void);
 int rd_wilderness_2(void);
 int rd_wilderness_3(void);
+int rd_wilderness_4(void);
 int rd_wilderness(void);
 int rd_dungeon(void);
 int rd_chunks(void);