+++ src/ui-map.h
@@ -24,3 +24,5 @@ extern void prt_map(void);
 extern void display_map(int *cy, int *cx);
 extern void do_cmd_view_map(void);
 void do_cmd_view_world_map(void);
+void ui_enter_magetower(game_event_type type, game_event_data *data,
+						void *user);